backend/pyproject.toml
[project]
name = "crypto-ai-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.104.1",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "matplotlib>=3.10.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
]
